now, expect debug directives are also tested after 'hlt' instruction that comes after 'ret'

+++ build-tests-isa.sh
@@ -1,15 +1,25 @@
 #!/bin/bash
 
 log=
+test_files=()
+
 for arg in "$@"; do
     if [[ $arg == -log-* ]]; then
         log="$arg"
-        break
+    else
+        test_files+=("$arg")
     fi
 done
 
+echo "Test files: ${test_files[*]}"
 
 for a in tests/isa/*.asm; do
+    fname=$(basename "$a" .asm)
+    # Skip if test_files is not empty and the current file is not in the list
+    if [ ${#test_files[@]} -ne 0 ] && [[ ! " ${test_files[@]} " =~ " $fname " ]]; then
+        continue
+    fi
+
     b=`echo $a | sed 's/\.asm/.bin/'`
     l=`echo $a | sed 's/\.asm/.dis/'`
     rm -rf $b
@@ -22,4 +32,11 @@ for a in tests/isa/*.asm; do
     ndisasm -b 16 $b > $l
 done
 
-cargo r --bin emu8086 -- -test -hide-header $log -show-binary-name -dump-regs-on-halt tests/isa/*.bin
+# Adjust the cargo command to use only the selected test files
+if [ ${#test_files[@]} -eq 0 ]; then
+    test_binaries="tests/isa/*.bin"
+else
+    test_binaries=$(printf "tests/isa/%s.bin " "${test_files[@]}")
+fi
+
+cargo r --bin emu8086 -- -test -hide-header $log -show-binary-name -dump-regs-on-halt $test_binaries
